Module raw_window_handle::borrowed

Borrowable window handles based on the ones in this crate.

These should be 100% safe to pass around and use, no possibility of dangling or invalidity.

Structs

DisplayHandle

The handle to the display controller of the windowing system.

WindowHandle

The handle to a window.

Traits

HasDisplayHandle

A display that acts as a wrapper around a display handle.

HasWindowHandle

A handle to a window.