Crate raw_window_handle

Interoperability library for Rust Windowing applications.

This library provides standard types for accessing a window’s platform-specific raw window handle. This does not provide any utilities for creating and managing windows; instead, it provides a common interface that window creation libraries (e.g. Winit, SDL) can use to easily talk with graphics libraries (e.g. gfx-hal).

Safety guarantees

Please see the docs of HasRawWindowHandle.

Platform handle initialization

Each platform handle struct is purposefully non-exhaustive, so that additional fields may be added without breaking backwards compatibility. Each struct provides an empty method that may be used along with the struct update syntax to construct it. See each specific struct for examples.

Structs

AndroidNdkHandle

Raw window handle for Android NDK.

AppKitHandle

Raw window handle for AppKit.

HaikuHandle

Raw window handle for Haiku.

OrbitalHandle

Raw window handle for the Redox operating system.

UiKitHandle

Raw window handle for UIKit.

WaylandHandle

Raw window handle for Wayland.

WebHandle

Raw window handle for the Web.

Win32Handle

Raw window handle for Win32.

WinRtHandle

Raw window handle for WinRT.

XcbHandle

Raw window handle for Xcb.

XlibHandle

Raw window handle for Xlib.

Enums

RawWindowHandle

An enum to simply combine the different possible raw window handle variants.

Traits

HasRawWindowHandle

Window that wraps around a raw window handle.