Enum raw_window_handle::RawWindowHandle

#[non_exhaustive]
pub enum RawWindowHandle {

    UiKit(UiKitWindowHandle),
    AppKit(AppKitWindowHandle),
    Orbital(OrbitalWindowHandle),
    OhosNdk(OhosNdkWindowHandle),
    Xlib(XlibWindowHandle),
    Xcb(XcbWindowHandle),
    Wayland(WaylandWindowHandle),
    Drm(DrmWindowHandle),
    Gbm(GbmWindowHandle),
    Win32(Win32WindowHandle),
    WinRt(WinRtWindowHandle),
    Web(WebWindowHandle),
    WebCanvas(WebCanvasWindowHandle),
    WebOffscreenCanvas(WebOffscreenCanvasWindowHandle),
    AndroidNdk(AndroidNdkWindowHandle),
    Haiku(HaikuWindowHandle),

}

A window handle for a particular windowing system.

Each variant contains a struct with fields specific to that windowing system (e.g. Win32WindowHandle will include a HWND, WaylandWindowHandle uses wl_surface, etc.)

Variant Availability

Note that all variants are present on all targets (none are disabled behind #[cfg]s), but see the “Availability Hints” section on each variant for some hints on where this variant might be expected.

Note that these “Availability Hints” are not normative. That is to say, a HasWindowHandle implementor is completely allowed to return something unexpected. (For example, it’s legal for someone to return a RawWindowHandle::Xlib on macOS, it would just be weird, and probably requires something like XQuartz be used).

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UiKit(UiKitWindowHandle)

A raw window handle for UIKit (Apple’s non-macOS windowing library).

Availability Hints

This variant is likely to be used on iOS, tvOS, (in theory) watchOS, and Mac Catalyst ($arch-apple-ios-macabi targets, which can notably use UIKit or AppKit), as these are the targets that (currently) support UIKit.

AppKit(AppKitWindowHandle)

A raw window handle for AppKit.

Availability Hints

This variant is likely to be used on macOS, although Mac Catalyst ($arch-apple-ios-macabi targets, which can notably use UIKit or AppKit) can also use it despite being target_os = "ios".

Orbital(OrbitalWindowHandle)

A raw window handle for the Redox operating system.

Availability Hints

This variant is used by the Orbital Windowing System in the Redox operating system.

OhosNdk(OhosNdkWindowHandle)

A raw window handle for the OpenHarmony OS NDK

Availability Hints

This variant is used on OpenHarmony OS (target_env = "ohos").

Xlib(XlibWindowHandle)

A raw window handle for Xlib.

Availability Hints

This variant is likely to show up anywhere someone manages to get X11 working that Xlib can be built for, which is to say, most (but not all) Unix systems.

Xcb(XcbWindowHandle)

A raw window handle for Xcb.

Availability Hints

This variant is likely to show up anywhere someone manages to get X11 working that XCB can be built for, which is to say, most (but not all) Unix systems.

Wayland(WaylandWindowHandle)

A raw window handle for Wayland.

Availability Hints

This variant should be expected anywhere Wayland works, which is currently some subset of unix systems.

Drm(DrmWindowHandle)

A raw window handle for the Linux Kernel Mode Set/Direct Rendering Manager

Availability Hints

This variant is used on Linux when neither X nor Wayland are available

Gbm(GbmWindowHandle)

A raw window handle for the Linux Generic Buffer Manager.

Availability Hints

This variant is present regardless of windowing backend and likely to be used with EGL_MESA_platform_gbm or EGL_KHR_platform_gbm.

Win32(Win32WindowHandle)

A raw window handle for Win32.

Availability Hints

This variant is used on Windows systems.

WinRt(WinRtWindowHandle)

A raw window handle for WinRT.

Availability Hints

This variant is used on Windows systems.

Web(WebWindowHandle)

A raw window handle for the Web.

Availability Hints

This variant is used on Wasm or asm.js targets when targeting the Web/HTML5.

WebCanvas(WebCanvasWindowHandle)

A raw window handle for a Web canvas registered via wasm-bindgen.

Availability Hints

This variant is used on Wasm or asm.js targets when targeting the Web/HTML5.

WebOffscreenCanvas(WebOffscreenCanvasWindowHandle)

A raw window handle for a Web offscreen canvas registered via wasm-bindgen.

Availability Hints

This variant is used on Wasm or asm.js targets when targeting the Web/HTML5.

AndroidNdk(AndroidNdkWindowHandle)

A raw window handle for Android NDK.

Availability Hints

This variant is used on Android targets.

Haiku(HaikuWindowHandle)

A raw window handle for Haiku.

Availability Hints

This variant is used on HaikuOS.